Develop custom get methods, everything working

+++ src/main/java/co/com/chartsofka/music/dto/SongDTO.java
@@ -14,7 +14,7 @@ public class SongDTO {
     private String songID;
     private String name;
     private Integer duration;
-    private Integer played;
+    private Long played;
     private AlbumDTO albumDTO;
